+++ client/hooks/useTheme.tsx
@@ -1,6 +1,7 @@
 import React, { useCallback, useContext, useEffect, useState } from "react";
 
 export enum Theme {
+  AUTO = "auto",
   LIGHT = "light",
   DARK = "dark",
 }
@@ -25,13 +26,17 @@ export function ThemeProvider({ init = Theme.LIGHT, children }: ThemeProviderPro
   const onThemeChange = useCallback(async (newTheme: Theme) => {
     document.documentElement.classList.remove(...Object.values(Theme));
     document.documentElement.classList.add(newTheme);
+    setCookie("theme", newTheme);
     setTheme(newTheme);
   }, []);
   
-  // Refreshes the cookie
+  // Refresh the cookie
   useEffect(() => {
-    setCookie("theme", theme);
-  }, [theme]);
+    const inCookie = getCookie("theme");
+    if(inCookie) {
+      setCookie("theme", inCookie);
+    }
+  }, []);
   
   return <ThemeContext.Provider value={[theme, onThemeChange]}>{children}</ThemeContext.Provider>;
 }
@@ -41,4 +46,11 @@ function setCookie(name: string, value: string) {
   const expires = new Date();
   expires.setFullYear(expires.getFullYear() + 1);
   document.cookie = `${name}=${value}; expires=${expires.toUTCString()}; path=/; SameSite=Lax`;
+}
+
+function getCookie(name: string) {
+  const value = `; ${document.cookie}`;
+  const parts = value.split(`; ${name}=`);
+  if(parts.length === 2) return parts.pop()!.split(';').shift();
+  else return "";
 }